chore(pcb-engine): slice 2 wrap-up
Fix a clippy cast warning in lint, tick plan task 6, record slice-2
findings in the spec: rip-up engages only on asymmetric congestion
(prospective cost load-balances symmetric layouts first-pass),
structurally-impassable vs congested edge distinction, FailedNet move,
InvalidLayer lint closing the slice-1 blind spot.

+++ crates/pcb-engine/src/lint.rs
@@ -149,7 +149,7 @@ pub enum DrcViolation {
 pub fn lint(problem: &RouteProblem, solution: &RouteSolution) -> Vec<DrcViolation> {
     let items = collect_items(problem, solution);
     let clearance = problem.clearance;
-    let layer_count = problem.layer_count.max(1) as u32;
+    let layer_count = problem.layer_count.max(1);
 
     let mut out = Vec::new();
